// vacupet_server/src/Enums/QuestionTypes.ts
export enum QuestionTypes {
  TEXT = "877f5354-dbc2-40d8-9fe5-39e84c210572",
  DATE = "11d72178-f5a7-4d7a-a099-88cfc1ba1645",
  NUMERIC = "3c8201ae-3519-45be-8d89-871f5f111d2b",
  STATES = "51b1992f-5c17-4294-a552-bb63f43c5704",
  BANKS = "c3de47c2-7a0d-4146-b5a9-aadf7b42dec1",
  CHECKBOX = "0a8bf07f-f66c-4d5f-a4aa-39267563f188",
  JUSTCHECK = "4cc0db57-4e43-48f5-aaaa-4ef7365b0589",
}